refactor undo keylistener
